Refactor: 코스용 페이지 사용을 위해 리뷰리스트 컴포넌트로 분리
